requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0
click>=8.1.0
jsonschema>=4.0.0
Pillow>=10.0.0

# KiCad integration
# Note: pcbnew module comes with KiCad installation
# For cross-version compatibility:
# kicad-python>=0.4.0  # Optional: for newer IPC-based API

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0